consolidated Product and Location models
