+++ src/main/java/com/revature/vanqapp/model/GroceryList.java
@@ -2,11 +2,7 @@ package com.revature.vanqapp.model;
 
 import com.fasterxml.jackson.annotation.*;
 import javax.persistence.*;
-import com.revature.vanqapp.model.product.Product;
 import lombok.*;
-
-import java.util.HashSet;
-import java.util.Set;
 
 @Entity
 @Data
