+++ src/main/java/com/revature/vanqapp/model/Location.java
@@ -3,6 +3,8 @@ package com.revature.vanqapp.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
+import javax.persistence.Table;
+
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Location {
@@ -10,4 +12,15 @@ public class Location {
     private StoreAddress address;
     private String phone;
     private String name;
+
+    @Data
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @Table(name = "address")
+    public static class StoreAddress {
+        private String addressLine1;
+        private String city;
+        private String state;
+        private String zipCode;
+        private String county;
+    }
 }
