--- src/main/java/com/revature/vanqapp/model/product/Item.java
@@ -1,13 +1,0 @@
-package com.revature.vanqapp.model.product;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Data;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Item {
-    private String itemId;
-    private Price price;
-    private String size;
-}
-
