+++ src/main/java/com/revature/vanqapp/repository/ProductRepository.java
@@ -1,10 +1,9 @@
 package com.revature.vanqapp.repository;
 
-import com.revature.vanqapp.model.product.Product;
-import com.revature.vanqapp.model.product.ProductLocationId;
+import com.revature.vanqapp.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, ProductLocationId> {
+public interface ProductRepository extends JpaRepository<Product, Product.ProductLocationId> {
 }
