+++ src/main/java/com/revature/vanqapp/service/ProductDeserializer.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import com.revature.vanqapp.model.product.Product;
+import com.revature.vanqapp.model.Product;
 
 import java.io.IOException;
