+++ src/main/java/com/revature/vanqapp/service/ProductService.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.revature.vanqapp.model.AuthToken;
-import com.revature.vanqapp.model.product.ProductFilterTerms;
-import com.revature.vanqapp.model.product.Product;
+import com.revature.vanqapp.model.ProductFilterTerms;
+import com.revature.vanqapp.model.Product;
 import com.revature.vanqapp.repository.KrogerApiRepository;
 import com.revature.vanqapp.util.AuthTokenFactoryBean;
 import org.apache.commons.pool2.ObjectPool;
